Moved GetUniqueHash to seperate class
