+++ conwaylib/Board.cs
@@ -2,10 +2,6 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using System.Security.Cryptography;
-using System.IO;
-using System.Runtime.Serialization;
-using System.Runtime.Serialization.Formatters.Binary;
 
 namespace ConwayLib
 {
@@ -14,7 +10,6 @@ namespace ConwayLib
   /// initialiser syntax by virtue of implementing <see cref="IEnumerable{T}"/> and having the
   /// <see cref="Add"/> method.
   /// </summary>
-  [Serializable]
   public sealed class Board : IMutableBoard, IEnumerable<bool>
   {
     private readonly bool[][] mCells;
@@ -28,17 +23,8 @@ namespace ConwayLib
 
     public byte[] GetUniqueHash()
     {
-      using(HashAlgorithm sha = SHA256.Create())
-      {
-        using (var stream = new MemoryStream())
-        {
-          IFormatter formatter = new BinaryFormatter(); 
-          formatter.Serialize(stream, mCells);
-          var rawBytes = stream.ToArray();
-          stream.Close();
-          return sha.ComputeHash(rawBytes);
-        }
-      }
+      var hasher = new BoolArrayHasher();
+      return hasher.GetUniqueHash(mCells.SelectMany(row => row));
     } 
 
     public int Width => mCells.FirstOrDefault()?.Length ?? 0;
